backendRAG/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
distro==1.9.0
fastapi==0.115.12
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.72.0
packaging==24.2
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
reportlab==4.3.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
